Remove SearchUsersByName usecase
