+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/ChatDomain.kt
@@ -489,25 +489,6 @@ public sealed interface ChatDomain {
     public fun deleteChannel(cid: String): Call<Unit>
 
     /**
-     * Perform api request with a search string as autocomplete if in online state. Otherwise performs search by name
-     * in local database.
-     *
-     * @param querySearch Search string used as autocomplete.
-     * @param offset Offset for paginated requests.
-     * @param userLimit The page size in the request.
-     * @param userPresence Presence flag to obtain additional info such as last active date.
-     *
-     * @return Executable async [Call] querying users.
-     */
-    @CheckResult
-    public fun searchUsersByName(
-        querySearch: String,
-        offset: Int,
-        userLimit: Int,
-        userPresence: Boolean,
-    ): Call<List<User>>
-
-    /**
      * Query members of a channel.
      *
      * @param cid CID of the Channel whose members we are querying.
